feat: Post 페이지 내 PurchaseButton 컴포넌트 구현
