+++ breaking-front/src/pages/Post/Post.js
@@ -3,20 +3,15 @@ import { useNavigate, useParams } from 'react-router-dom';
 import dayjs from 'dayjs';
 import { UserInformationContext } from 'providers/UserInformationProvider';
 import useInfiniteScroll from 'hooks/useInfiniteScroll';
-import usePostBoughtList from 'pages/Post/hooks/queries/usePostBoughtList';
 import usePost from 'hooks/queries/usePost';
 import usePostComment from 'pages/Post/hooks/queries/usePostComment';
 import usePostLike from 'pages/Post/hooks/mutations/usePostLike';
 import useDeletePostLike from 'pages/Post/hooks/mutations/useDeletePostLike';
-import usePostBuy from 'pages/Post/hooks/mutations/usePostBuy';
 import { PAGE_PATH } from 'constants/path';
 import { PostSkeleton } from 'components/Skeleton/Skeleton';
 import ScrollToTop from 'components/ScrollToTop/ScrollToTop';
 import Tag from 'components/Tag/Tag';
 import Line from 'components/Line/Line';
-import Button from 'components/Button/Button';
-import Modal from 'components/Modal/Modal';
-import FollowCard from 'components/FollowCard/FollowCard';
 import ContentSlice from 'components/ContentSlice/ContentSlice';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
 import InfiniteTargetDiv from 'components/InfiniteTargetDiv/InfiniteTargetDiv';
@@ -24,6 +19,7 @@ import Carousel from 'pages/Post/components/Carousel/Carousel';
 import Comment from 'pages/Post/components/Comment/Comment';
 import ContentToggle from 'pages/Post/components/ContentToggle/ContentToggle';
 import CommentForm from 'pages/Post/components/CommentForm/CommentForm';
+import PurchaseButton from 'pages/Post/components/PurchaseButton/PurchaseButton';
 import ImageUrlConverter from 'utils/ImageUrlConverter';
 import * as Style from 'pages/Post/Post.styles';
 import { ReactComponent as LocationIcon } from 'assets/svg/location.svg';
@@ -38,8 +34,6 @@ const Post = () => {
   const navigate = useNavigate();
   const { userId, profileImgURL } = useContext(UserInformationContext);
 
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [purchaseType, setPurchaseType] = useState('구매하기');
   const [isOpenContentToggle, setIsOpenContentToggle] = useState(false);
 
   const [isLiked, setIsLiked] = useState(false);
@@ -47,8 +41,7 @@ const Post = () => {
   const [likeCount, setLikeCount] = useState(0);
 
   const { data: postData, isLoading: isPostDataLoading } = usePost(postId);
-  const { data: postBoughtList, refetch: PostBoughtListReFetch } =
-    usePostBoughtList(postId);
+
   const {
     data: postCommentData,
     isFetching: isPostCommentFetching,
@@ -63,21 +56,8 @@ const Post = () => {
   const { mutate: PostLike } = usePostLike();
   const { mutate: DeletePostLike } = useDeletePostLike();
 
-  const { mutate: PostBuy } = usePostBuy();
-
-  const toggleButListModal = () => {
-    PostBoughtListReFetch();
-    setIsModalOpen((pre) => !pre);
-  };
-
   const profileClick = () => {
     navigate(PAGE_PATH.PROFILE(postData?.data.user?.userId));
-  };
-
-  const postBuyClick = () => {
-    const postBuyConfirm = window.confirm('게시글을 구매하시겠습니까?');
-
-    postBuyConfirm && PostBuy(postId);
   };
 
   const toggleLiked = () => {
@@ -91,11 +71,6 @@ const Post = () => {
   };
 
   useEffect(() => {
-    if (postData?.data.isMyPost) setPurchaseType('구매자 목록');
-    else if (postData?.data.isPurchased) setPurchaseType('다운로드');
-    else if (!postData?.data.isPurchasable) setPurchaseType('판매 중지');
-    else setPurchaseType('구매 하기');
-
     setIsLiked(postData?.data.isLiked);
     setLikeCount(postData?.data.likeCount);
   }, [postData, userId]);
@@ -103,19 +78,6 @@ const Post = () => {
   return (
     <>
       <ScrollToTop />
-      <Modal
-        isOpen={isModalOpen}
-        closeClick={toggleButListModal}
-        title="구매자 목록"
-      >
-        {postBoughtList?.data.map((item) => (
-          <FollowCard
-            cardClick={() => navigate(PAGE_PATH.PROFILE(item.userId))}
-            profileData={item}
-            key={item.userId}
-          />
-        ))}
-      </Modal>
       <Style.Post>
         {isPostDataLoading ? (
           <PostSkeleton />
@@ -179,25 +141,12 @@ const Post = () => {
                 <Style.ContentPrice>
                   {postData?.data.price.toLocaleString('ko-KR')}&nbsp;원
                 </Style.ContentPrice>
-                {purchaseType === '구매자 목록' && (
-                  <Button color="primary" onClick={toggleButListModal}>
-                    {purchaseType}
-                  </Button>
-                )}
-                {purchaseType === '구매 하기' && (
-                  <Button color="primary" onClick={postBuyClick}>
-                    {purchaseType}
-                  </Button>
-                )}
-                {purchaseType === '판매 중지' && (
-                  <Button color="secondary" disabled>
-                    {purchaseType}
-                  </Button>
-                )}
-                {purchaseType === '다운로드' && (
-                  //다운로드 로직 추후 구현
-                  <Button color="primary">{purchaseType}</Button>
-                )}
+                <PurchaseButton
+                  postId={postId}
+                  isMyPost={postData?.data.isMyPost}
+                  isPurchased={postData?.data.isPurchased}
+                  isPurchasable={postData?.data.isPurchasable}
+                />
                 <Style.ContentDetail>
                   누적 판매
                   <Style.ContentSoldCount>
